Reduce size of TaskId to 4 bytes

+++ crates/tako/src/lib.rs
@@ -9,7 +9,7 @@ pub mod transfer;
 pub mod worker;
 
 define_id_type!(WorkerId, u32);
-define_id_type!(TaskId, u64);
+define_id_type!(TaskId, u32);
 define_id_type!(InstanceId, u32);
 
 // Priority: Bigger number -> Higher priority
